requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
python-jose
passlib[bcrypt]
openai
pdfplumber
fpdf